// src/auth/auth.controller.interface.ts
import { Request, Response, NextFunction } from 'express';

export interface IAuthController {
  login: (req: Request, res: Response, next: NextFunction) => void;
  register: (req: Request, res: Response, next: NextFunction) => void;
  info: (req: Request, res: Response, next: NextFunction) => void;
}